* dev/remove-windows-warnings:
  Remove unused variable.
  Add casting to remove type conversion warnings.
  Prevent symbol from being defined twice.
  Ad malloc include file to quell malloc/free warnings.

+++ vital/bindings/c/helpers/c_utils.cxx
@@ -42,7 +42,7 @@ namespace vital_c {
 void make_string_list( std::vector<std::string> const &list,
                        unsigned int &length, char ** &strings )
 {
-  length = list.size();
+  length = static_cast< unsigned int >(list.size());
   strings = (char**)malloc( sizeof(char*) * length );
   for( unsigned int i = 0; i < length; i++ )
   {
